backend/requirements.txt
django
djangorestframework
django-cors-headers
django-environ
django-filter
django-redis
django-storages
boto3
stripe
pillow
psycopg2-binary
djangorestframework-simplejwt
drf-nested-routers
daphne
channels
channels-redis
mammoth
whitenoise
pdfplumber
dj-database-url
gunicorn
cloudinary
django-cloudinary-storage
